Add files via upload
This is the first version which really works.  Includes highlighted (inverted) digit on programming.  This version has a major change: in the previous versions I did all processing and display writing in the receiveEvent interrupt but I found this blocked the second set of data during programming (the original blink alternate pixel page) so moved the display processing to the main loop which runs continuously and asynchronously to the receiveEvent routine, which has been kept as short as possible.